+++ Frontend/Parser/Parser.h
@@ -27,7 +27,7 @@ namespace Coda {
 			Token expect(TokenType type, std::string error);
 			Node parseStatement();
 			Node parseExpression();
-			Node parseFunctionExpression();
+			Node parseFunctionExpression(std::string name);
 			Node parseAssignmentExpression();
 			Node parseBlockExpression();
 			Node parseObjectExpression();